Make DFT2 more robust to large N limit.

+++ LPSD/genwin.h
@@ -11,8 +11,8 @@ void set_window (int type, double req_psll, char *name, double *psll,
 
 void makewinsincos (long int nfft, double bin, double *win, double *winsum,
 		            double *winsum2, double *nenbw);
-void makewinsincos_indexed (long int nfft, double bin, double *win, double *winsum,
-		            double *winsum2, double *nenbw, int, int, bool);
+void makewinsincos_indexed (unsigned long int nfft, double bin, double *win, double *winsum,
+		            double *winsum2, double *nenbw, unsigned long int, unsigned int, bool);
 
 void makewin (unsigned int nfft, double *win,
               double *winsum, double *winsum2, double *nenbw);
